requirements.txt
Flask
Flask-CORS
PyJWT
Werkzeug
bcrypt
pymysql
